refactor(dependencies): introduce separate module for dependencies
