+++ lib/convenient_service/dependencies.rb
@@ -1,0 +1,4 @@
+# frozen_string_literal: true
+
+require_relative "dependencies/built_in"
+require_relative "dependencies/extractions"
