+++ lib/convenient_service/dependencies/built_in.rb
@@ -1,0 +1,33 @@
+# frozen_string_literal: true
+
+##
+# This file load built-in dependencies (from Ruby standard library).
+#
+
+##
+# @internal
+#   - https://ruby-doc.org/stdlib-2.7.0/libdoc/forwardable/rdoc/Forwardable.html
+#   - https://github.com/ruby/forwardable
+#
+require "forwardable"
+
+##
+# @internal
+#   - https://ruby-doc.org/stdlib-2.7.0/libdoc/logger/rdoc/Logger.html
+#   - https://github.com/ruby/logger
+#
+require "logger"
+
+##
+# @internal
+#   - https://ruby-doc.org/stdlib-2.7.0/libdoc/ostruct/rdoc/OpenStruct.html
+#   - https://github.com/ruby/ostruct
+#
+require "ostruct"
+
+##
+# @internal
+#   - https://ruby-doc.org/stdlib-2.7.0/libdoc/singleton/rdoc/Singleton.html
+#   - https://github.com/ruby/singleton
+#
+require "singleton"
