+++ lib/convenient_service/dependencies/extractions/active_support_concern.rb
