+++ lib/convenient_service/dependencies/extractions/byebug_syntax_highlighting.rb
