+++ lib/convenient_service/dependencies/extractions/ruby_middleware.rb
@@ -3,7 +3,7 @@
 ##
 # NOTE:
 #   Copied from `Ibsciss/ruby-middleware` without any logic modification. Version: v0.4.2.
-#   Just wrapped in a namespace `ConvenientService::Extractions::RubyMiddleware`.
+#   Just wrapped in a namespace `ConvenientService::Dependencies::Extractions::RubyMiddleware`.
 #   Replaced `require` into `require_relative`.
 #
 # - https://github.com/marian13/ruby-middleware/blob/v0.4.2/middleware.gemspec
