+++ lib/convenient_service/dependencies/extractions/ruby_middleware/middleware.rb
