+++ lib/convenient_service/support/concern.rb
@@ -4,7 +4,7 @@ module ConvenientService
   module Support
     module Concern
       def self.included(klass)
-        klass.extend Extractions::ActiveSupportConcern::Concern
+        klass.extend Dependencies::Extractions::ActiveSupportConcern::Concern
       end
     end
   end
